feat(renderer): one unified diff renderer — jsdiff fallback, shared with conflict view
Step 3 of the code-editor workstream (spec §6).

- DiffView + helpers extracted from ToolBody to components/diff/UnifiedDiff;
  the structuredPatch path (absolute file line numbers) is byte-identical, the
  hand-rolled O(m*n) LCS fallback is replaced by jsdiff (declared dependency
  that nothing in the renderer imported)
- EOF-newline normalization keeps the old LCS semantics: an EOF-newline-only
  change is noise at row granularity
- The artifact conflict 'View diff' now renders a real unified diff (was two
  raw <pre> columns) labeled as what 'Keep mine' would change on disk
- Preserved deliberately: DIFF_PREVIEW_LINES capping, hunk separators, calc()
  gutter, hardcoded red/green (theme-INDEPENDENT by design)
